[BSE-919] Use the memory budgets in join/groupby (#6367)
* Generate code at the start/end of the function with operator comptroller calls

* [run CI]

* docstrings

* updated generated test output

* Allow adding OperatorComptroller finalize without a return

* Update docstrings

* Change test import

* Fix StreamingStateScopeTest (account for compute_satisfiable_budgets call)

* Add tests for new CodegenFrame methods

* Add tests for new StreamingStateScope functionality

* [run CI]

* Update test outputs

* [run CI]

* [run CI]

* [run CI]

* Only generate operatorComptroller calls when streaming

* [run CI]

* Fix pipeline counting

* [run CI]

* [run CI]

* [run CI]

* Only check last element for addBeforeReturn

* Add a few TODOs

* [run CI]

* Fix typo

* [run CI]

* [run CI]

* [run CI]

* [run CI]

* Add tests of compute_satisfiable_budgets

* [run CI]

* [run CI]

* Add operator_id to init_table

* Add operator_id to init_groupby_state

* Add operator_id to init_join_state

* Add operator_id to snowflake_writer_init

* Add operator_id to init_dict_encoding_state

* [run CI]

* Fix test code

* updated expectations

* [run CI]

* Fix incorrect generated code

* Fix stream join tests

* Fix stream dict encoding tests

* Fix stream groupby tests

* [run CI]

* Fix test_stream_hash_join_repartitioning

* fix test_table_builder

* support operator_id in union

* Fix test_union

* [run CI]

* [run CI]

* Use memory budget in join

* Use memory budget in group by

* Allow operator_id = -1 in GetOperatorBudget

This avoids a bit of boilerplate in every usage.

* [run CI]

* Disable memory budgeting by default for now

* Check for null in flag

* [run CI]

* Request all memory if estimate is -1

* Fix disabling budgeting

* Update tests to use before_each/after_each

* [run CI]

* [run CI]

* [run CI]

* [run CI]

+++ bodo/libs/_memory_budget.h
@@ -49,6 +49,8 @@ class OperatorComptroller {
     /**
      * @brief Get the budget for a given operator in the current pipeline
      * @param operator_id The operator owning the budget returned
+     * If the operator_id is invalid (< 0), returns -1 to signify unlimited
+     * budget
      */
     int64_t GetOperatorBudget(int64_t operator_id);
 
@@ -80,7 +82,7 @@ class OperatorComptroller {
 
     std::vector<size_t> pipeline_remaining_budget;
     std::vector<size_t> pipeline_remaining_operators;
-    std::vector<size_t> operator_allocated_budget;
+    std::vector<int64_t> operator_allocated_budget;
 
     struct OperatorRequest {
         int64_t min_pipeline_id;
